requirements.txt
requests
Pillow
opencv-python
colormath
MIDIUtil
torch
torchvision
numpy
tqdm
scikit-image